requirements.txt
langchain-community
langchain-core
python-dotenv

langgraph
fastapi
uvicorn[standard]
pydantic
requests
python-multipart
langchain-openai
pandas
